Convert Advert.spec to TS
